+++ metadata.rb
@@ -1,5 +1,5 @@
 name    'etcd'
-version '2.1.1'
+version '2.1.2'
 
 depends 'ark'
 depends 'partial_search', '~> 1.0.6'