+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * fix kitchen test on cluster to pickup this issue.
 * add basic chefspec as well to ensure this sort of issue doesn't occur
 * fix bug with name attribute methods
+* fix searching for wrong recipe when using cluster recipe
 
 ### 2.1.1
 * update to etcd 0.3.0